Bug 486947: Use count objects returned by PrototypeServiceFactory
For OSGi R6 the framework does not use count the 'same' objects that get
returned by a PrototypeServiceFactory.  When a bundle calls
ServiceObjects.getService() multiple times a PrototypeServiceFactory may
return the 'same' object multiple times.  The bundle then may call
ServiceObjects.ungetService(S) with the 'same' service object multiple
times.  Currently the framework implementation will throw an
IllegalArgumentException on the second call to ungetService(S) if the
'same' service object is passed in more than once.  This is because we
do not use count the service object for ServiceObjects.getService()

The core expert group has decided the framework must use count the
service objects returned by ServiceObjects.getService() and decrement
the use count on calls to ServiceObjects.ungetService(S).